+++ R/getSubCategory.R
@@ -6,25 +6,44 @@
 #' @param onlySid2 sid2 is sub category id. some sub categories don't have id. If TRUE, functions return data.frame(chr:sub_cate_naem, char:sid2). Defaults is TRUE.
 #' @return Get data.frame(chr:sub_cate_name, chr:urls).
 #' @export
-#' @import xml2
-#' @import rvest
-#' @import stringr
+#' @importFrom rvest html_nodes html_attr html_text
+#' @importFrom httr GET content user_agent
 
-getSubCategory <- function(sid1="100", onlySid2=TRUE) {
+getSubCategory <- function(sid1 = 100, onlySid2 = TRUE) {
+  root <-
+    paste0("http://news.naver.com/main/main.nhn?mode=LSD&mid=shm&sid1=",
+           sid1)
+  uat <-
+    httr::user_agent("N2H4 by chanyub.park <mrchypark@gmail.com>")
+  src <- httr::GET(root, uat)
+  hobj <- httr::content(src)
+  titles <- rvest::html_nodes(hobj, "div.snb ul.nav li a")
+  titles <- rvest::html_text(titles)
+  titles <- trimws(titles)
 
-  print("This function use internet. If get error, please check the internet.")
-  home <- paste0("http://news.naver.com/main/main.nhn?mode=LSD&mid=shm&sid1=",sid1)
-  tem <- read_html(home)
-  titles <- tem %>% html_nodes("div.snb ul.nav li a") %>% html_text()
-  titles<-str_trim(titles)
-  links <- tem %>% html_nodes("div.snb ul.nav li a") %>% html_attr("href")
-  links <- paste0("http://news.naver.com",links)
-  urls <- data.frame(sub_cate_name=titles,url=links,stringsAsFactors = F)
-  if (onlySid2==FALSE)  {return(urls)}
+  links <- rvest::html_nodes(hobj, "div.snb ul.nav li a")
+  links <- rvest::html_attr(links, "href")
+  links <- paste0("http://news.naver.com", links)
+
+  urls <-
+    data.frame(
+      sub_cate_name = titles,
+      url = links,
+      stringsAsFactors = F
+    )
+  if (onlySid2 == FALSE)  {
+    return(urls)
+  }
   else{
-    urls <- urls[grep("sid2=",urls$url),]
-    sid2 <- str_sub(urls$url,-3,-1)
-    urls <- data.frame(sub_cate_name=urls$sub_cate_name,sid2=sid2,stringsAsFactors = F)
+    urls <- urls[grep("sid2=", urls$url),]
+    sid2 <- sapply(strsplit(urls$url, "="), function(x)
+      x[5])
+    urls <-
+      data.frame(
+        sub_cate_name = urls$sub_cate_name,
+        sid2 = sid2,
+        stringsAsFactors = F
+      )
     return(urls)
   }