+++ R/getUrlListByCategory.R
@@ -6,31 +6,43 @@
 #' @param  col is what you want to get from news. Defualt is all.
 #' @return Get data.frame(titles, links).
 #' @export
-#' @import xml2
-#' @import rvest
-#' @import stringr
+#' @importFrom rvest html_nodes html_attr html_text
+#' @importFrom httr GET content user_agent
 
-getUrlListByCategory <- function(turl = url, col=c("titles", "links")) {
+getUrlListByCategory <-
+  function(turl = url,
+           col = c("titles", "links")) {
+    uat <-
+      httr::user_agent("N2H4 by chanyub.park <mrchypark@gmail.com>")
+    src <- httr::GET(turl, uat)
+    hobj <- httr::content(src)
 
-    tem <- xml2::read_html(turl)
-    titles <- tem %>% rvest::html_nodes("dt a") %>% rvest::html_text()
+    titles <- rvest::html_nodes(hobj, "dt a")
+    titles <- rvest::html_text(titles)
     Encoding(titles) <- "UTF-8"
-    rm_target <- tem %>% rvest::html_nodes("dt.photo a") %>% rvest::html_text()
+
+    rm_target <- rvest::html_nodes(hobj, "dt.photo a")
+    rm_target <- rvest::html_text(rm_target)
     Encoding(rm_target) <- "UTF-8"
-    links <- tem %>% rvest::html_nodes("dt a") %>% rvest::html_attr("href")
 
-    news_lists <- data.frame(titles = titles, links = links, stringsAsFactors = F)
+    links <- rvest::html_nodes(hobj, "dt a")
+    links <- rvest::html_attr(links, "href")
 
-    news_lists$titles <- str_trim(news_lists$titles, side="both")
-    news_lists <- news_lists[nchar(news_lists$titles) > 0,]
+    news_lists <-
+      data.frame(titles = titles,
+                 links = links,
+                 stringsAsFactors = F)
 
-    rm_target <- str_trim(rm_target, side="both")
+    news_lists$titles <- trimws(news_lists$titles)
+    news_lists <- news_lists[nchar(news_lists$titles) > 0, ]
+
+    rm_target <- trimws(rm_target)
     rm_target <- rm_target[nchar(rm_target) > 0]
 
     if (!identical(paste0(rm_target, collapse = " "), "")) {
-      news_lists <- news_lists[-grep(rm_target[1], news_lists$titles),]
+      news_lists <- news_lists[-grep(rm_target[1], news_lists$titles), ]
     }
 
-    return(news_lists[,col])
+    return(news_lists[, col])
 
-}
+  }