+++ tests/testthat/test-func.R
@@ -3,7 +3,7 @@ context("function test")
 test_that("getComment", {
 
   url  <- "http://news.naver.com/main/read.nhn?mode=LSD&mid=shm&sid1=102&oid=001&aid=0009205077"
-  test <- getComment(url)
+  test <- getComment(url, type="list")
   test <- test$result$commentList[[1]]
   expect_equal(test$contents, "test")
 
@@ -44,7 +44,7 @@ test_that("getMaxPageNum", {
 
   url  <- "http://news.naver.com/main/list.nhn?sid2=254&sid1=102&mid=shm&mode=LS2D&date=20170427"
   test <- getMaxPageNum(url)
-  expect_equal(test, 2)
+  expect_equal(test, 1)
 
 })